Jupiter is not the 9th planet 😉 (#1475)
If the reference to "after Mars," is
desired for some reason, punctuation should
probably be added for clarity. E.g., "planet,
after Mars,", or "planet, following Mars,". Note
also that because this is a FIM example and "the
5th" is part of the sentence to be filled, it
shouldn't be included as part of the completion.

continuedev commit: f3d200afe46fe7a9ff938997301e60af8148db69