Refresh index fixes (#1795)
* Use the resultType given to markComplete

This fixes an issue where if there were multiple result types in itemToAction such as an update with UpdateNewVersion and UpdateOldVersion, only the UpdateOldVersion would be hit (which is a noop) and the UpdateNewVersion would never be fired, causing us to never update the timestamp of the tag_catalog.

* Add a Compute and UpdateLastUpdated to refreshIndex

Removes the ability to insert duplicate rows if we're just computing the result. Instead we replace the existing row instead.

* Don't insert compute tags

They are REPLACEd in markComplete now.

continuedev commit: 447e1e686d071b4ae235f35f4a8b42b32266d57e