Change the LanceDBIndex to compute its new cache rows in parallel and then do a bulk insert (#1944)
This greatly improves the performance of indexing. Combined with the
change for bulk inserts in ChunkCodebaseIndexer, indexing is over twice
as fast. However, it can be much faster if the other indexers are
converted to use bulk insert as well.

In addition, this indexer is not using the computed chunks in the chunk
table. It would be faster to reference the already computed chunks,
however, for the purposes of this PR we are not changing that behavior.

Co-authored-by: Rob Leidle <rleidle@tesla.com>

continuedev commit: 88c72c342673d74d55cacd2b7a38119e6ad089b1