Reliably trigger VSCode extension rebuild
The code to trigger a rebuild on missing extension.js added in
https://github.com/continuedev/continue/pull/3276
turns out not to be entirely reliable because esbuild sometimes
falls back to full-contents comparison rather than using the modtime.
To fix this, include a buildTimestamp file in the input, and
rewrite that when we want to trigger a rebuild.

continuedev commit: 2483abdc7b498d02d6c0405662f7c70890a4f33b