When inserting streaming diffs, don't mess up for blank lines
There was code of unclear purpose that bailed out early when
totalDeletedContent == "" - but "" could either be

 [].join("\n")

or:

 [""].join("\n")

in the latter case, we *do* need to a deletion block. Since the
deletionBuffer.length == 0 case is otherwise handled, the code
is removed as unnecessary and harmful.

continuedev commit: 4219a206f2407d170cf75d7abb6a86e0561bc4b8