Fixed multiline selection issue in edit mode
The original behavior was only selecting the highlighted part
and if there is nothing selected in the trailing line it still
got seleced to the edit mode. This fix is to expand the
selection start to the very beginning of first line and the
selection end to the very end of the last line. In addtion,
prevent the trailing line getting selecting if there is
nothing in it.
https://github.com/Granite-Code/granite-code/issues/42

continuedev commit: 742ca318ceb505790fa6962cf4b70073cb0a5c07