Added contextLength attribute to LLMOptions which was missing
The contextLength attribute exists under both the completionOptions and
defaultCompletionOptions attributes, but not at the root level. As a
result, the reference in core/llm/llms/Ollama.ts is undefined. To fix
this, we simply copied the contextLength from defaultCompletionOptions to
the root attribute of LLMOptions

continuedev commit: 32105490b9156eff5828722060015df518d11286