Fix OpenRouter models response validation
preferredIndex is not always present in the OpenRouter response, which means that zod will raise an error.

Nullish in zod terms means that an item is both optional and nullable, which resolves the validation error.

Note: This change also resolved a bug for me in the SettingsView that caused it to detect changes any time I opened it. However, going back to a clean `main` I can't replicate the issue any longer, so I'm not 100% sure if this fixed it. (It feels like maybe the cache was in a bad state which kept triggering the changes, and somehow this change fixed that?)