feat(vscode): add a vscode task to watch and run pnpm install
We currently have three watch tasks that auto run when vs code launches. The problem is if the packages changes, some of these tasks will start to silently fail. Now, we've added a new `pnpm:watch` task that will watch the pnpm-lockfile for changes and auto run `pnpm i` when it changes. The other tasks now depend on the `pnpm` task now.

Hopefully this will make switching branches easier without having to manually mess with the tasks.