feat: Report issues to Sentry
Included tags:
* ide_version
* plugin_version
* jcef_supported

To test it, just register an action that throws an exception. As a result, a notification appears + there is a blinking icon in the bottom-right corner. Click on it and select 'Submit to Continue'. User message is also sent to Sentry (you can leave it empty).

continuedev commit: 3b82a30382c54796f4e46a6bd1cc988cc94b0ee4