Fix: add sanitizing process to strip off the trailing slash of the doc URL
In this patch, we add a sanitizing process to strip off the trailing forward slash (if any) in
the given URL before generating the file path for S3 bucket. Therefore, both variants (url
with / and without /) in the config file can be treated the same.

https://github.com/continuedev/continue/issues/6454

continuedev commit: f59863ba70793a98460056d967f982ca457a6315