test(webview-ui): set NODE_ENV to test in vitest setup
We have some kilocode changes in the `SettingsView` that we're trying to reset the settings to the original state and that code is breaking the new way that the settings auto approval toggles are being set.

This code:
```
	// From time to time there's a bug that triggers unsaved changes upon rendering the SettingsView
	// This is a (nasty) workaround to detect when this happens, and to force overwrite the unsaved changes
	const renderStart = useRef<null | number>()
	useEffect(() => {
		renderStart.current = performance.now()
	}, [])
	useEffect(() => {
		if (renderStart.current && process.env.NODE_ENV !== "test") {
			const renderEnd = performance.now()
			const renderTime = renderEnd - renderStart.current

			if (renderTime < 100 && isChangeDetected) {
				console.info("Overwriting unsaved changes in less than 100ms")
				onConfirmDialogResult(true)
			}
		}
	}, [isChangeDetected, onConfirmDialogResult])
	// kilocode_change end
```

which should only run if the `NODE_ENV` is set to `test`. But it looks like in the `vitest.setup.ts`, they explicitly set the `NODE_ENV` to `development`, which seems odd. So I set it back to `test`, which makes it so our code does not kick in, which fixes the test failures in the `SettingsView.spec.tsx`!