fix: fix JCEF browser issues (#7291)
* fix: fix JCEF browser issues

Solves CON-2469
Solves CON-2979

Off-screen rendering is now enabled by default. This is motivated by the fact that most users will benefit from it,
but the minority with issues can disable it in the Continue settings. I believe this is the correct strategy,
because disabling it causes a flickering window for the majority.

Note: I tried to mask it by preloading the HTML with a background color and by changing the background color of the
browser component. That worked, but only reduced the flashing time and didn't completely eliminate the problem.
Enabling Off-screen rendering seems to be the only robust solution.

After decoupling the browser from the toolbar and moving it to a dedicated service, there is now full control over
when and what is initialized. The setting is now applied correctly.

I've noticed that most of the actions require a working, active Continue toolbar.
I've created the `ContinueToolbarAction` abstraction to remove duplicated logic. I've also removed `actions.utils`.

* Fix bug in `CoreMessenger.handleMessage` (some pass-through messages had message type used as message ID)
* Add `browser` package and moved browser-related classes there, including `CustomHandlerFactory` (package `factories` is now removed). `CustomHandlerFactory` implementation is untouched.
* Add `project.getBrowser()` extension method (this is mostly for readability).
* Some loggers now use a shorter identifier (simple class name) without the full package, because our package name is
  unusually long, and it was hard to read the logs.
* Remove commented-out `TerminalActivityTrackingService`.
* Create `JcefErrorPanel` and put `jcef_error.html` there.

* fix: enable OSR by default + remove from settings

* fix: remove ApplyToFileHandlerTest.kt

* fix: attempt to fix CI error + remove outdated JCEF registry options

continuedev commit: 62f468abf949ec248cdb8dacb50f1d9f1082a51c